+++ app/components/MainSection.jsx
@@ -1,12 +1,12 @@
-import React, { Component, PropTypes } from 'react';
+import React, { PropTypes } from 'react';
 import TopicItem from 'components/TopicItem';
 import classNames from 'classnames/bind';
-import styles from 'scss/components/_main-section';
+import styles from 'scss/components/main-section';
 
 const cx = classNames.bind(styles);
 
 const MainSection = ({onIncrement, onDecrement, onDestroy, topics}) => {
-  
+
   const topicItems = topics.map((topic, key) => {
     return (
       <TopicItem index={key}