+++ app/containers/Navigation.jsx
@@ -1,17 +1,20 @@
-import React, { Component, PropTypes } from 'react';
+import React, { PropTypes } from 'react';
 import { Link } from 'react-router';
 import { connect } from 'react-redux';
 import { logOut } from 'actions/users';
-import styles from 'scss/components/_navigation';
 
-class Navigation extends Component {
+import classNames from 'classnames/bind';
+import styles from 'scss/components/navigation';
 
-  render() {
-    const { dispatch } = this.props;
+const cx = classNames.bind(styles);
+
+const Navigation = ({user, dispatch}) => {
     return (
       <nav className={styles.navigation} role="navigation">
-          <Link to="/" className={styles.navigation__item + ' ' + styles['navigation__item--logo']} activeClassName={styles['navigation__item--active']}>Ninja Ocean</Link>
-          { this.props.user.authenticated ? (
+        <Link to="/"
+          className={cx('navigation__item', 'navigation__item--logo')}
+          activeClassName={styles['navigation__item--active']}>Ninja Ocean</Link>
+          { user.authenticated ? (
             <Link onClick={()=> dispatch(logOut())}
               className={styles.navigation__item} to="/">Logout</Link>
           ) : (
@@ -21,9 +24,7 @@ class Navigation extends Component {
           <Link to="/about" className={styles.navigation__item} activeClassName={styles['navigation__item--active']}>About</Link>
       </nav>
     );
-  }
-
-}
+};
 
 Navigation.propTypes = {
   user: PropTypes.object,