+++ app/reducers/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import user from 'reducers/user';
 import topic from 'reducers/topic';
+import message from 'reducers/message';
 import { routeReducer as routing } from 'react-router-redux';
 
 // Combine reducers with routeReducer which keeps track of
@@ -8,6 +9,7 @@ import { routeReducer as routing } from 'react-router-redux';
 const rootReducer = combineReducers({
   user,
   topic,
+  message,
   routing
 });