+++ app/reducers/user.js
@@ -1,4 +1,5 @@
 import {
+  TOGGLE_LOGIN_MODE,
   MANUAL_LOGIN_USER,
   LOGIN_SUCCESS_USER,
   LOGIN_ERROR_USER,
@@ -7,29 +8,40 @@ import {
   SIGNUP_ERROR_USER,
   LOGOUT_USER,
   LOGOUT_SUCCESS_USER,
-  LOGOUT_ERROR_USER } from 'constants';
+  LOGOUT_ERROR_USER } from 'constants/index';
 
 export default function user(state={
+  isLogin: true,
+  message: '',
   isWaiting: false,
   authenticated: false }, action={}) {
   switch (action.type) {
+    case TOGGLE_LOGIN_MODE:
+      return Object.assign({}, state, {
+        isLogin: !state.isLogin,
+        message: ''
+      });
     case MANUAL_LOGIN_USER:
       return Object.assign({}, state, {
-        isWaiting: true
+        isWaiting: true,
+        message: ''
       });
     case LOGIN_SUCCESS_USER:
       return Object.assign({}, state, {
         isWaiting: false,
-        authenticated: true
+        authenticated: true,
+        message: ''
       });
     case LOGIN_ERROR_USER:
       return Object.assign({}, state, {
         isWaiting: false,
-        authenticated: false
+        authenticated: false,
+        message: action.message
       });
     case SIGNUP_USER:
       return Object.assign({}, state, {
-        isWaiting: true
+        isWaiting: true,
+        message: ''
       });
     case SIGNUP_SUCCESS_USER:
       return Object.assign({}, state, {
@@ -39,11 +51,13 @@ export default function user(state={
     case SIGNUP_ERROR_USER:
       return Object.assign({}, state, {
         isWaiting: false,
-        authenticated: false
+        authenticated: false,
+        message: action.message
       });
     case LOGOUT_USER:
       return Object.assign({}, state, {
-        isWaiting: true
+        isWaiting: true,
+        message: ''
       });
     case LOGOUT_SUCCESS_USER:
       return Object.assign({}, state, {
@@ -53,7 +67,8 @@ export default function user(state={
     case LOGOUT_ERROR_USER:
       return Object.assign({}, state, {
         isWaiting: false,
-        authenticated: true
+        authenticated: true,
+        isLogin: true
       });
     default:
       return state;