+++ app/routes.jsx
@@ -4,7 +4,7 @@ import { Route, IndexRoute } from 'react-router';
 import App from 'containers/App';
 import Vote from 'containers/Vote';
 import About from 'containers/About';
-import Login from 'containers/Login';
+import LoginOrRegister from 'containers/LoginOrRegister';
 import Dashboard from 'containers/Dashboard';
 
 /*
@@ -23,10 +23,20 @@ export default (store) => {
     }
     callback();
   };
+
+  const redirectAuth = (nextState, replace, callback) => {
+    const { user: { authenticated }} = store.getState();
+    if (authenticated) {
+      replace({
+        pathname: '/'
+      });
+    }
+    callback();
+  };
   return (
     <Route path="/" component={App}>
       <IndexRoute component={Vote} />
-      <Route path="login" component={Login} />
+      <Route path="login" component={LoginOrRegister} onEnter={redirectAuth} />
       <Route path="dashboard" component={Dashboard} onEnter={requireAuth} />
       <Route path="about" component={About} />
     </Route>