+++ server/controllers/users.js
@@ -10,21 +10,25 @@ exports.postLogin = function(req, res, next) {
   passport.authenticate('local', function(err, user, info) {
     if(err) return next(err);
     if(!user) {
-      req.flash('errors', {msg: info.message});
+     return res.status(401).json({ message: info.message});
     }
-    // Passport exposes a login() function on req (also aliased as logIn()) that can be used to establish a login session
+    // Passport exposes a login() function on req (also aliased as
+    // logIn()) that can be used to establish a login session
     req.logIn(user, function(err) {
-      if(err) return next(err);
-      res.status(200).send('OK');
+      if(err) return res.status(401).json({message: err});
+      return res.status(200).json(
+        {
+          message: 'You have been successfully logged in.'
+        });
     });
   })(req, res, next);
 };
 
 
 /**
- * GET /logout
+ * POST /logout
  */
-exports.getLogout = function(req, res) {
+exports.postLogout = function(req, res) {
   // Do email and password validation for the server
   req.logout();
   res.redirect('/');
@@ -42,14 +46,16 @@ exports.postSignUp = function(req, res, next) {
 
   User.findOne({email: req.body.email}, function(err, existingUser) {
     if(existingUser) {
-      req.flash('errors', { msg: 'Account with that email address already exists' });
+      return res.status(409).json({ message: 'Account with this email address already exists!'});
     }
     user.save(function(err) {
       if(err) return next(err);
       req.logIn(user, function(err) {
-        if(err) return next(err);
-        console.log('Successfully created');
-        res.end('Success');
+        if(err) return res.status(401).json({message: err});
+        return res.status(200).json(
+          {
+            message: 'You have been successfully logged in.'
+          });
       });
     });
   });